% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{source_R}
\alias{source_R}
\title{source_R}
\usage{
source_R(dir = getwd(), recursive = TRUE)
}
\arguments{
\item{dir}{dir to search in. Defaults to working directory.}

\item{recursive.}{bool. passed to list.files}
}
\description{
Return list of R scripts
}
