% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEM_mediation.R
\name{lavaan_mediation}
\alias{lavaan_mediation}
\title{lavaan_mediation}
\usage{
lavaan_mediation(latents, outcome, direct, mediator)
}
\arguments{
\item{latents}{named list, name defines latent name, vector defines contents}

\item{outcome}{string, outcome variable}

\item{direct}{string, variable that forms the direct path}

\item{mediator}{string, variable that forms the mediator path}
}
\description{
lavaan_mediation
}
\details{
https://lavaan.ugent.be/tutorial/mediation.html
}
