% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snipe.R
\name{snipe}
\alias{snipe}
\title{snipe}
\usage{
snipe(path = getwd(), dir_name = "Analysis", manuscript = NULL)
}
\arguments{
\item{path}{path to data object}

\item{dir_name}{name of desired top-level folder}

\item{manuscript}{name of the manuscript file if desired, or TRUE to include}
}
\description{
Initialise targets around a data object. The original data object will be moved to inside the new targets structure.
}
