% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_bignetwork.R
\name{tar_bignetwork}
\alias{tar_bignetwork}
\title{tar_bignetwork}
\usage{
tar_bignetwork(width = "1020px", height = "980px")
}
\arguments{
\item{width}{string including "px"}

\item{height}{string including "px"}
}
\description{
Makes a bigger html widget tar_visnetwork
}
