% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsl_code_bat.R
\name{wsl_nvim_bat}
\alias{wsl_nvim_bat}
\title{wsl_nvim_bat}
\usage{
wsl_nvim_bat(path)
}
\arguments{
\item{path}{Path to the Windows directory to open in WSL and launch Neovim.}
}
\description{
Creates a .bat file that opens a Windows directory in WSL (mounted) and launches Neovim in that directory.
}
