% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsl_code_bat.R
\name{wsl_vscode_bat}
\alias{wsl_vscode_bat}
\title{wsl_vscode_bat}
\usage{
wsl_vscode_bat(path)
}
\arguments{
\item{path}{Path to the Windows directory to open in WSL and launch Visual Studio Code.}
}
\description{
Creates a .bat file that opens a Windows directory in WSL (mounted) and launches Visual Studio Code in that directory.
}
