% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_R_scripts}
\alias{get_R_scripts}
\title{get_R_scripts}
\usage{
get_R_scripts(dir = getwd(), recursive = TRUE)
}
\arguments{
\item{dir}{dir to search in. Defaults to working directory.}

\item{recursive.}{bool. passed to list.files}
}
\description{
Return list of R scripts
}
