% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{n_workers}
\alias{n_workers}
\title{n_workers}
\usage{
n_workers(
  reserve_n = NULL,
  reserve_p = 0.6,
  max_workers = Inf,
  ncpus = Sys.getenv("snipe_ncpus")
)
}
\arguments{
\item{reserve_n}{numeric. How many workers to reserve for other operations. If NULL, reserves a third of workers}

\item{reserve_p}{numeric. Proportion of workers to reserve for other operations. Used when reserve_n is NULL}

\item{max_workers}{numeric. Maximum number of workers to return}

\item{ncpus}{numeric. Number of workers to use, can be set in .Renviron using snipe_ncpus}
}
\description{
Automatically select number of workers to use in parallel operations
}
