% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{n_workers}
\alias{n_workers}
\title{n_workers}
\usage{
n_workers(x = 10)
}
\arguments{
\item{x}{numeric. How many workers to reserve for other operations}
}
\description{
Automatically select number of workers
}
