% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_tar_data.R
\name{snipe}
\alias{snipe}
\title{snipe}
\usage{
snipe(path, dir_name = "Analysis", manuscript = NULL)
}
\arguments{
\item{path}{path to data object}

\item{dir_name}{name of desired top-level folder}

\item{manuscript}{name of the manuscript file if desired, or TRUE to include}
}
\description{
Initialise targets around a data object
}
